requirements.txt
Flask~=1.1.2
requests~=2.24.0
PyYAML~=5.3.1

APScheduler~=3.6.3
